+++ src/screens/CardList/CardListView.tsx
@@ -5,9 +5,9 @@ import commonStyles from '../../styles/commonStyles';
 import colors from '../../styles/Colors';
 import ScreenContainer from '../../components/ScreenContainer';
 import SearchInput from '../../components/cardList/SearchInput';
-import SortDropDown from '../../components/cardList/SortDropDown';
-import CardSectionList from '../../components/cardList/CardSectionList';
-import CardBottomSheet from '../../components/cardList/CardBottomSheet';
+import SortDropDown from '../../components/cardList/elements/SortDropDown';
+import CardSectionList from '../../components/cardList/sectionListView/SectionList';
+import CardBottomSheet from '../../components/cardList/elements/CardBottomSheet';
 import EmptyListView from '../../components/cardList/EmptyListView';
 import { useCardStore } from '../../store/cardStore';
 import { dummyData } from '../../model/cardItem';
